Tests/LibWeb/Text/expected/UIEvents/click-event-on-input-control-associated-with-label-element.txt

mouseup event on box
click event on box
click event on input#radio1
click event on input#radio2
